Misc name corrections (some are disputable depending on which leak you would prefer to reference) Placed all constants on the left side of conditional statements Removed redundant final modifiers on methods in final classes Standardized field/method order BuildEnvironment -> ModeWhere ClientPacket -> ClientProt Removed unnecessary qualifiers Replace some loop copies with System.arraycopy (intrinsic) Replaced magic constants such as those related to the Calendar class's months. Added missing @Override annotations So much varied cleanup I could never mention it all

+++ src/com/jagex/Class142.java
@@ -9,8 +9,8 @@ public class Class142 {
         double d_0 = 3.834951969714103E-4D;
 
         for (int i_2 = 0; i_2 < 16384; i_2++) {
-            aFloatArray1666[i_2] = (float) Math.sin((double) i_2 * d_0);
-            aFloatArray1665[i_2] = (float) Math.cos((double) i_2 * d_0);
+            aFloatArray1666[i_2] = (float) Math.sin(i_2 * d_0);
+            aFloatArray1665[i_2] = (float) Math.cos(i_2 * d_0);
         }
 
     }
