+++ src/com/jagex/Class297.java
@@ -9,8 +9,8 @@ public class Class297 {
         double d_0 = 3.834951969714103E-4D;
 
         for (int i_2 = 0; i_2 < 16384; i_2++) {
-            aFloatArray3537[i_2] = (float) Math.sin((double) i_2 * d_0);
-            aFloatArray3536[i_2] = (float) Math.cos((double) i_2 * d_0);
+            aFloatArray3537[i_2] = (float) Math.sin(i_2 * d_0);
+            aFloatArray3536[i_2] = (float) Math.cos(i_2 * d_0);
         }
 
     }
